requirements.txt
streamlit
chromadb
langchain
langchain-community
langchain-groq
sentence-transformers
python-dotenv
pdfplumber
pdfminer.six